Update RasterizeRDD Methods
